Completely remove deprecated bootloaders from clr-boot-manager
gummiboot hasn't been supported upstream or by Clear Linux for an age,
and due to format bumps no compatibility is required with it, thus it
can be completely removed from source.

goofiboot, the fork of gummiboot, was initially created to address concerns
in systemd-boot in their tooling. However, the concerns were mitigated by
opting to entirely bypass their tooling, and clr-boot-manager absorbed all
safety and distribution support that was part of libnica and goofiboot.

At this point, both of the legacy bootloaders are safe to be nuked from
orbit and are no longer used in CBM-using distributions.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>