Do not fail an update on efivar creation failure
When there is an issue with efivars do not fail the entire update
operation. Since the kernel and bootloader can still be successfully
updated and the bios can manually be updated to use the installed
bootloader, the efivar update is non-fatal. The operation being
treated as fatal could prevent update content from getting provisioned
to the EFI partition which is worse than not setting the bios
correctly for booting the updated system.

As part of this change add an error message to warn users they may
need to update their bios to select their desired boot option.