Enable commandline removal configuration files
Add the ability for clr-boot-manager to handle removing content from
the kernel commandline. The removal configuration files are formatted
exactly the same as the add configuration files. The matches are done
line-by-line so that:

"option-a option-b"

in a removal file would only delete exactly that string unlike:
"option-a
option-b"

in a removal file which would try to remove "option-a" and "option-b"
separately.