Attempt kernel update even if bootloader update fails
In some cases the bootloader update will not impact the kernel update,
usually when another tool like syslinux is used which may fail
independently of the file copies the kernel update will need. Since
the kernel update may still succeed always attempt it to avoid cases
where the kernel becomes stale as it is more likely to need updates
than the bootloader.