bootman: fix boot directory for "partitionless boot" - native_mode
When running into a system with a partition table where /boot hasn't
its own partition and running in native_mode we must also check if
/boot is empty to determine if we must keep trying to mount or not.

In native mode and without /boot partition we'll always fail to
detect boot device (once we don't have one) but will always report
a fail. This patch makes sure to check if in native we have content
installed into /boot directory and assume it's a partitionless /boot.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>