bootman: only do partitionless boot when the bootloader supports it
#174 restored support for partitionless /boot. However, this
introduced issues when (for whatever reason) there is content in
the /boot directory (instead of being empty) (see #175).

As partitionless boot is not an option for every bootloader, add a flag
indication bootloader support. Bootloaders without partitionless support
(and those on UEFI) will never try to use a partitionless /boot.